Added the %dopar% operator
Fixed a bug where I missed the %dopar% operator before the inner loop.

Also removed the .combine=rbind, because the assignment is being performed inside the parallel loop. The assignment is not a thread safe write, but the code guarantees that only one element is written at a time, so there's no reason to think that the wrong value will be assigned to each element in the matrix.